+++ src/app/rtl_connector.hpp
@@ -26,7 +26,7 @@ class RtlConnector: public Connector {
         virtual int set_center_frequency(double frequency) override;
         virtual int set_sample_rate(double sample_rate) override;
         virtual int set_gain(GainSpec* gain) override;
-        virtual int set_ppm(int32_t ppm) override;
+        virtual int set_ppm(double ppm) override;
         int set_direct_sampling(int direct_sampling);
 #if HAS_RTLSDR_SET_BIAS_TEE
         int set_bias_tee(bool bias_tee);
